got parent filter to work
